+++ internal/storagenode/executor/writer_test.go
@@ -14,6 +14,7 @@ import (
 	"go.uber.org/goleak"
 
 	"github.com/kakao/varlog/internal/storagenode/storage"
+	"github.com/kakao/varlog/internal/storagenode/telemetry"
 	"github.com/kakao/varlog/pkg/types"
 	"github.com/kakao/varlog/pkg/util/syncutil/atomicutil"
 	"github.com/kakao/varlog/pkg/verrors"
@@ -83,7 +84,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		committer:  committer,
 		replicator: replicator,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -95,7 +96,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		committer:  committer,
 		replicator: replicator,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -107,7 +108,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		committer:  committer,
 		replicator: replicator,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -119,7 +120,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		committer:  committer,
 		replicator: replicator,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -131,7 +132,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		lsc:        lsc,
 		replicator: replicator,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -143,7 +144,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		lsc:       lsc,
 		committer: committer,
 		state:     state,
-		me:        NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:   telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 
@@ -155,7 +156,7 @@ func TestWriterInvalidArgument(t *testing.T) {
 		lsc:        lsc,
 		committer:  committer,
 		replicator: replicator,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.Error(t, err)
 }
@@ -186,7 +187,7 @@ func TestWriterStop(t *testing.T) {
 		committer:  testCommitter.mock,
 		replicator: testReplicator.mock,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.NoError(t, err)
 
@@ -319,7 +320,7 @@ func TestWriter(t *testing.T) {
 			committer:  testCommitter.mock,
 			replicator: testReplicator.mock,
 			state:      state,
-			me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+			metrics:    telemetry.NewMetrics(),
 		})
 		require.NoError(t, err)
 
@@ -405,7 +406,7 @@ func TestWriterCleanup(t *testing.T) {
 		committer:  testCommitter.mock,
 		replicator: testReplicator.mock,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.NoError(t, err)
 
@@ -462,7 +463,7 @@ func TestWriterVarlog444(t *testing.T) {
 		committer:  testCommitter,
 		replicator: testReplicator.mock,
 		state:      state,
-		me:         NewTestMeasurableExecutor(ctrl, 1, 1),
+		metrics:    telemetry.NewMetrics(),
 	})
 	require.NoError(t, err)